+++ packages/rum-core/src/domain/rumEventsCollection/view/trackViewMetrics.ts
@@ -1,5 +1,5 @@
-import type { Duration, RelativeTime, Observable } from '@datadog/browser-core'
-import { noop, round, ONE_SECOND } from '@datadog/browser-core'
+import type { Duration, RelativeTime, Observable, ClocksState } from '@datadog/browser-core'
+import { noop, round, ONE_SECOND, elapsed } from '@datadog/browser-core'
 import type { RumLayoutShiftTiming } from '../../../browser/performanceCollection'
 import { supportPerformanceTimingEvent } from '../../../browser/performanceCollection'
 import { ViewLoadingType } from '../../../rawRumEvent.types'
@@ -19,7 +19,8 @@ export function trackViewMetrics(
   lifeCycle: LifeCycle,
   domMutationObservable: Observable<void>,
   scheduleViewUpdate: () => void,
-  loadingType: ViewLoadingType
+  loadingType: ViewLoadingType,
+  viewStart: ClocksState
 ) {
   const viewMetrics: ViewMetrics = {
     eventCounts: {
@@ -34,15 +35,15 @@ export function trackViewMetrics(
     scheduleViewUpdate()
   })
 
-  const { setActivityLoadingTime, setLoadEvent } = trackLoadingTime(loadingType, (newLoadingTime) => {
-    viewMetrics.loadingTime = newLoadingTime
-    scheduleViewUpdate()
-  })
-
-  const { stop: stopActivityLoadingTimeTracking } = trackActivityLoadingTime(
+  const { stop: stopLoadingTimeTracking, setLoadEvent } = trackLoadingTime(
     lifeCycle,
     domMutationObservable,
-    setActivityLoadingTime
+    loadingType,
+    viewStart,
+    (newLoadingTime) => {
+      viewMetrics.loadingTime = newLoadingTime
+      scheduleViewUpdate()
+    }
   )
 
   let stopCLSTracking: () => void
@@ -58,7 +59,7 @@ export function trackViewMetrics(
   return {
     stop: () => {
       stopEventCountsTracking()
-      stopActivityLoadingTimeTracking()
+      stopLoadingTimeTracking()
       stopCLSTracking()
     },
     setLoadEvent,
@@ -66,7 +67,13 @@ export function trackViewMetrics(
   }
 }
 
-function trackLoadingTime(loadType: ViewLoadingType, callback: (loadingTime: Duration) => void) {
+function trackLoadingTime(
+  lifeCycle: LifeCycle,
+  domMutationObservable: Observable<void>,
+  loadType: ViewLoadingType,
+  viewStart: ClocksState,
+  callback: (loadingTime: Duration) => void
+) {
   let isWaitingForLoadEvent = loadType === ViewLoadingType.INITIAL_LOAD
   let isWaitingForActivityLoadingTime = true
   const loadingTimeCandidates: Duration[] = []
@@ -77,7 +84,18 @@ function trackLoadingTime(loadType: ViewLoadingType, callback: (loadingTime: Dur
     }
   }
 
+  const { stop } = waitIdlePage(lifeCycle, domMutationObservable, (event) => {
+    if (isWaitingForActivityLoadingTime) {
+      isWaitingForActivityLoadingTime = false
+      if (event.hadActivity) {
+        loadingTimeCandidates.push(elapsed(viewStart.timeStamp, event.end))
+      }
+      invokeCallbackIfAllCandidatesAreReceived()
+    }
+  })
+
   return {
+    stop,
     setLoadEvent: (loadEvent: Duration) => {
       if (isWaitingForLoadEvent) {
         isWaitingForLoadEvent = false
@@ -85,30 +103,7 @@ function trackLoadingTime(loadType: ViewLoadingType, callback: (loadingTime: Dur
         invokeCallbackIfAllCandidatesAreReceived()
       }
     },
-    setActivityLoadingTime: (activityLoadingTime: Duration | undefined) => {
-      if (isWaitingForActivityLoadingTime) {
-        isWaitingForActivityLoadingTime = false
-        if (activityLoadingTime !== undefined) {
-          loadingTimeCandidates.push(activityLoadingTime)
-        }
-        invokeCallbackIfAllCandidatesAreReceived()
-      }
-    },
   }
-}
-
-function trackActivityLoadingTime(
-  lifeCycle: LifeCycle,
-  domMutationObservable: Observable<void>,
-  callback: (loadingTimeValue: Duration | undefined) => void
-) {
-  return waitIdlePage(lifeCycle, domMutationObservable, (event) => {
-    if (event.hadActivity) {
-      callback(event.duration)
-    } else {
-      callback(undefined)
-    }
-  })
 }
 
 /**